+++ src/app/clients/clients-view/custom-dialogs/view-signature-dialog/view-signature-dialog.component.ts
@@ -12,10 +12,9 @@ import { ClientsService } from 'app/clients/clients.service';
 @Component({
   selector: 'mifosx-view-signature-dialog',
   templateUrl: './view-signature-dialog.component.html',
-  styleUrls: ['./view-signature-dialog.component.scss']
+  styleUrls: ['./view-signature-dialog.component.scss'],
 })
 export class ViewSignatureDialogComponent implements OnInit {
-
   /** Id of client signature in documents */
   signatureId: any;
   /** Signature Image */
@@ -27,10 +26,12 @@ export class ViewSignatureDialogComponent implements OnInit {
    * @param {MatDialogRef} dialogRef Component reference to dialog.
    * @param {any} data Documents data
    */
-  constructor(public dialogRef: MatDialogRef<ViewSignatureDialogComponent>,
-              private clientsService: ClientsService,
-              private sanitizer: DomSanitizer,
-              @Inject(MAT_DIALOG_DATA) public data: { documents: any[], id: string }) {
+  constructor(
+    public dialogRef: MatDialogRef<ViewSignatureDialogComponent>,
+    private clientsService: ClientsService,
+    private sanitizer: DomSanitizer,
+    @Inject(MAT_DIALOG_DATA) public data: { documents: any[]; id: string }
+  ) {
     const signature = this.data.documents.find((document: any) => document.name === 'clientSignature') || {};
     this.signatureId = signature.id;
     this.clientId = this.data.id;
@@ -39,12 +40,10 @@ export class ViewSignatureDialogComponent implements OnInit {
 
   ngOnInit() {
     if (this.signatureId) {
-      this.clientsService.getClientSignatureImage(this.clientId, this.signatureId).subscribe(
-        (base64Image: any) => {
-          this.signatureImage = this.sanitizer.bypassSecurityTrustResourceUrl(base64Image);
-        }, (error: any) => {}
-      );
+      this.clientsService.downloadClientDocument(this.clientId, this.signatureId).subscribe((res) => {
+        const url = window.URL.createObjectURL(res);
+        this.signatureImage = this.sanitizer.bypassSecurityTrustUrl(url);
+      });
     }
   }
-
 }